1st controller

+++ resources/views/welcome.blade.php
@@ -12,7 +12,7 @@
         </div>
         <a href="/show/{{ $image->id }}" class="btn btn-info">View</a>
         <a href="/edit/{{ $image->id }}" class="btn btn-warning">Edit</a>
-        <a href="/delete/{{ $image->id }}" class="btn btn-danger">Delete</a>
+        <a href="/delete/{{ $image->id }}" onclick="return confirm('Are you sure?')" class="btn btn-danger">Delete</a>
       </div>
     @endforeach
   </div>
